Optimized reads
Total rewrite of the Read operation. There is now a segment LRU
cache instead of multiple readers. This has more memory overhead,
because up to four segments may be loaded in memory at one time,
but is a ton faster and allows for forward and reverse sequential
reads.

This commit also removes the Durability option in favor of a
simpiler NoSync option.